Move lib_fft dsl functions to separate fft.hpp
